requirements.txt
numpy==2.4.2
pillow==12.1.0
google-genai==1.62.0
python-dotenv==1.2.1
